The program run script is changed to addax.sh to remove the dependency on Python
For Linux, Unix and MacOS, you can use addax.sh to run a job without python runtime.
But Python is required yet in Windows.

Note: Under MacOS, the addax.sh script ONLY supports the short options, because built-in getopts lacks support for the long option, and gnu getopt is not shipped by default