[bugfix][writer][oracle] Fix ORA-01461 With MERGE INTO Statement And Clob Column #1030
ORA-1461 error occurs when executing merge statements with a NCLOB column as a target and a bind variable value which is longer than 4000 bytes:

Only if String length > 4000 bytes.

If instead of MERGE a separate statement is used for INSERT or UPDATE (based on situation) then the error is not reported.